src/test/java/br/gov/lexml/eta/etaservices/parsing/xml/EmendaXmlUnmarshallerTest.java
package br.gov.lexml.eta.etaservices.parsing.xml;

import br.gov.lexml.eta.etaservices.emenda.Emenda;
import org.dom4j.DocumentException;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class EmendaXmlUnmarshallerTest {

    private EmendaXmlUnmarshaller unmarshaller;

    @BeforeEach
    void setUp() {
        unmarshaller = new EmendaXmlUnmarshaller();
    }

    @Test
    void fromXml() throws DocumentException {
        Emenda e = null;
                //unmarshaller.fromXml("<emenda local=\"abc\" data=\"2022-08-10\"></emenda>");

        assertThat(e).isNull();

    }
}